% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDO_Heterotic2_Pvalue.r
\name{ADDO_Heterotic2_Pvalue}
\alias{ADDO_Heterotic2_Pvalue}
\title{Pvalue calculation and Verfication of Overdominance QTLs
       (STEP2 of The Heterotic Model)}
\usage{
ADDO_Heterotic2_Pvalue(indir = indir, outdir = outdir,
  Input_name = Input_name, Kinship_type = Kinship_type,
  VarComponent_Method = VarComponent_Method, PheList_Choose = F,
  PheList = PheList, Phe_HistogramPlot = F, Run_separated = F,
  covariates_sum = covariates_sum, Phe_IndMinimum = 200,
  GT_IndMinimum = 10, num_nodes = 10)
}
\arguments{
\item{indir}{A character. The input directory where contains the input bPLINK or GenABEL data.}

\item{outdir}{A character. The output directory where generates the folder: "2_Pvalue".}

\item{Input_name}{A character. The prefixes of the input files.}

\item{Kinship_type}{A character. The method to generate kinship matrix. Please select from "GenABEL","EMMA","EMMAX","GEMMA", "GCTA", "GCTA_ad", "HOMEBREW_AFW" or "HOMEBREW_AS".}

\item{VarComponent_Method}{A character. The method to estimate variance components. Please select from "EMMA_a", "GCTA_a" or "GCTA_ad" (When VarComponent_Method is "GCTA_ad", the Kinship_type must be "GCTA_ad").}

\item{PheList_Choose}{A logic variable. T: Just investigate specified phenotypes; F: Investigate all phenotypes.}

\item{PheList}{A vector of character. When choose "PheList_Choose=F", the specified phenotype list must be specified.}

\item{Phe_HistogramPlot}{A logic variable. T: Draw the histogram plots for all phenotypes; F: Aviod the histogram plots.}

\item{Run_separated}{A logic variable. T: Run the separated genotype files; F: Run the whole genotype file.}

\item{covariates_sum}{A numeric variable. The sum of all covariates.}

\item{Phe_IndMinimum}{A numeric variable. Remove phenotypes without enough available individuals.}

\item{GT_IndMinimum}{A numeric variable. Remove loci with available individuals <GT_IndMinimum for all three genotypes.}

\item{num_nodes}{A numeric variable. The number of cores used parallelly.}
}
\value{
a folder named "2_Pvalue" with various statistics of each significant SNP for all phenotypes.
}
\description{
P-value Calculation and Verfication of overdominance (or heterotic) QTLs
             (1) Run the whole PLINK file or Run the separated PLINK files ("Run_separated = F");
             (2) Indicate Reocde Model (AA:1 0 0/AB:0 1 0/BB:0 0 1) without the "1" column of covariance matrix;
             (3) Estimate two T-statistics (t(AB-AA) and t(AB-BB)) to measure the deviation between the effect of heterozygote (AB) and that of two homozygotes (AA and BB);
             (4) Generate the P-value based on MVN distribution using the minor(abs(t(AB-AA)),abs(t(AB-BB))) from SNPs with t(AB-AA)*t(AB-BB)>0.
}
\examples{
ADDO_Heterotic2_Pvalue(indir=indir, outdir=outdir, Input_name="TEST", Kinship_type="GCTA_ad", VarComponent_Method="GCTA_ad", covariates_sum=2)

}
\author{
Leilei Cui and Bin Yang
}
