% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDO_Heterotic3_Plot.r
\name{ADDO_Heterotic3_Plot}
\alias{ADDO_Heterotic3_Plot}
\title{Visulization of Various Different Figures
       (STEP3 of The Heterotic Model)}
\usage{
ADDO_Heterotic3_Plot(outdir = outdir, PheList_Choose = F,
  PheList = PheList, covariates_sum = covariates_sum,
  RegionMan_chr_whole = F, RegionMan_chr_region = RegionMan_chr_region,
  Down_sampling = F, Down_sampling_logP = 1,
  Down_sampling_distance = 10, chrs_sum = chrs_sum, num_nodes = 10)
}
\arguments{
\item{outdir}{A character. The output directory where generates the folder: "3_Plot".}

\item{PheList_Choose}{A logic variable. T: Just investigate specified phenotypes; F: Investigate all phenofiles.}

\item{PheList}{A vector of character. When choose "PheList_Choose=F", the specified phenotype list must be specified.}

\item{covariates_sum}{A numeric variable. The sum of all covariates.}

\item{RegionMan_chr_whole}{A logic variable. T: Draw a whole chromosome; F: Draw the specified region.}

\item{RegionMan_chr_region}{A numeric variable. The length of specified region around the Peak SNP.}

\item{Down_sampling}{A logic variable. T: Down-sampling points with low logP to speed up the plotting progress; F: Darwing with all loci.}

\item{Down_sampling_logP}{A numeric variable. The threshold for down-sampling points for rapid rendering of the Manhattan Plots, when Down_sampling is true.}

\item{Down_sampling_distance}{A numeric variable. The distance of points for equidistant sampling, when Down_sampling is true.}

\item{chrs_sum}{A numeric variable. The sum of all chromosomes.}

\item{num_nodes}{A numeric variable. The number of cores used parallelly.}
}
\value{
a folder named "3_Plot" with various plots.
}
\description{
Visualizing overdominance QTLs by various plots.
}
\examples{
ADDO_Heterotic3_Plot(outdir=outdir, covariates_sum=2, RegionMan_chr_whole=F, RegionMan_chr_region = 2000000)

}
\author{
Leilei Cui and Bin Yang
}
