% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGWAS.r
\name{plotGWAS}
\alias{plotGWAS}
\title{A Flexible Function to draw Manhattan Plot}
\usage{
plotGWAS(chrs = chrs, traitIdx = 1, plotcolor = c("darkgreen"),
  alldat = alldat, y_limit = "", main = "", cex_points = 1,
  cex_lab = 3.1, cex_axis = 3)
}
\arguments{
\item{chrs}{A vector of numbers or characters indicating the chromosomes of markers.}

\item{traitIdx}{A numeric variable indicating the index of trait to plot.}

\item{plotcolor}{A vector of characters specify the colors used to plot the signatures on each chromosome.}

\item{alldat}{A dataframe with columns containing SNP, chr, pos and association strength (-log10 P value) of markers for one or more traits.}

\item{y_limit}{A numeric variable detail the range of y axis.}

\item{main}{A character specifies the title of the plot.}

\item{cex_points}{A numeric value specifies the cex of points in the plot.}

\item{cex_lab}{A numeric value specifies the cex of labels in the plot.}

\item{cex_axis}{A numeric value specifies the cex of axis in the plot.}
}
\value{
a figure
}
\description{
Manhattan Plot for whole genome association analyses for single or multiple traits
}
\author{
Bin Yang and Leilei Cui
}
