% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.r
\name{plotRegion}
\alias{plotRegion}
\title{A Flexible Function for Manhattan Plot}
\usage{
plotRegion(chrs = chrs, traitIdx = 1, alldat = alldat, from = NULL,
  to = NULL, main = "", ldinfo = ldinfo, ylim = NULL,
  cex_points = 2.5, cex_points_peak = 4, cex_lab = 1.3,
  cex_axis = 1.3, cex_main = 1.4)
}
\arguments{
\item{chrs}{A vector of numbers or characters indicating the chromosomes of markers.}

\item{traitIdx}{A numeric variable indicating the index of trait to plot.}

\item{alldat}{A dataframe with columns containing SNP, chr, pos and association strength (-log10 P value) of markers for one or more traits.}

\item{from}{A character specify the name of SNP.}

\item{to}{A character specify the name of SNP.}

\item{main}{A character specifies the title of the plot.}

\item{ldinfo}{A data frame generated using --r2 --ld-snp command in PLINK .}

\item{ylim}{A vector of two numeric variables specify the range of y axis..}

\item{cex_points}{A numeric variable specify the cex of points to plot.}

\item{cex_points_peak}{A numeric variable specify the cex of point for the lmost significant marker to plot..}

\item{cex_lab}{A numeric value specifies the cex of labels in the plot.}

\item{cex_axis}{A numeric value specifies the cex of axis in the plot.}

\item{cex_main}{AA numeric value specifies the cex of title in the plot.}
}
\value{
a figure
}
\description{
Plot the GWAS results for a single trait on one particular chromosome, or specific region on one chromosome
}
\author{
Bin Yang and Leilei Cui
}
